Move common parts of cpu_info_x86 and cpu_info_arm classes into cpu_info_impl
