+++ cpp/oneapi/dal/detail/cpu_info_impl.cpp
@@ -14,7 +14,9 @@
 * limitations under the License.
 *******************************************************************************/
 
-#include "oneapi/dal/detail/cpu_info_iface.hpp"
+#include "oneapi/dal/detail/cpu_info_impl.hpp"
+
+#include <sstream>
 
 namespace oneapi::dal::detail {
 namespace v1 {
@@ -46,6 +48,44 @@ std::string to_string(cpu_extension extension) {
     return extension_str;
 }
 
+cpu_vendor cpu_info_impl::get_cpu_vendor() const {
+    return std::any_cast<detail::cpu_vendor>(info_.find("vendor")->second);
+}
+
+cpu_extension cpu_info_impl::get_top_cpu_extension() const {
+    return std::any_cast<cpu_extension>(info_.find("top_cpu_extension")->second);
+}
+
+std::string cpu_info_impl::dump() const {
+    std::stringstream ss;
+    for (auto it = info_.begin(); it != info_.end(); ++it) {
+        ss << it->first << " : ";
+        print_any(it->second, ss);
+        ss << "; ";
+    }
+    std::string result;
+    std::string token;
+    while (ss >> token) {
+        result += token + " ";
+    }
+    return result;
+}
+
+template <typename T>
+void cpu_info_impl::print(const std::any& value, std::stringstream& ss) const {
+    T typed_value = std::any_cast<T>(value);
+    ss << to_string(typed_value);
+}
+
+void cpu_info_impl::print_any(const std::any& value, std::stringstream& ss) const {
+    const std::type_info& ti = value.type();
+    if (ti == typeid(cpu_extension)) {
+        print<cpu_extension>(value, ss);
+    }
+    else if (ti == typeid(cpu_vendor)) {
+        print<cpu_vendor>(value, ss);
+    }
+}
+
 } // namespace v1
-using v1::to_string;
 } // namespace oneapi::dal::detail
